Move sidecar classes to their own package
