+++ wls-exporter-core/src/main/java/com/oracle/wls/exporter/AuthenticationChallengeException.java
@@ -1,4 +1,4 @@
-// Copyright (c) 2019, 2020, Oracle and/or its affiliates.
+// Copyright (c) 2019, 2021, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
 package com.oracle.wls.exporter;
@@ -6,10 +6,10 @@ package com.oracle.wls.exporter;
 /**
  * An exception indicating that the server has sent a challenge, demanding authentication.
  */
-class AuthenticationChallengeException extends WebClientException {
-    private String challenge;
+public class AuthenticationChallengeException extends WebClientException {
+    private final String challenge;
 
-    AuthenticationChallengeException(String challenge) {
+    public AuthenticationChallengeException(String challenge) {
         this.challenge = challenge;
     }
