+++ wls-exporter-core/src/main/java/com/oracle/wls/exporter/ForbiddenException.java
@@ -1,4 +1,4 @@
-// Copyright (c) 2017, 2020, Oracle and/or its affiliates.
+// Copyright (c) 2017, 2021, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
 package com.oracle.wls.exporter;
@@ -8,5 +8,5 @@ package com.oracle.wls.exporter;
  *
  * @author Russell Gold
  */
-class ForbiddenException extends WebClientException {
+public class ForbiddenException extends WebClientException {
 }
