+++ wls-exporter-sidecar/src/main/java/com/oracle/wls/exporter/sidecar/Main.java
@@ -1,10 +1,11 @@
 // Copyright (c) 2021, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package com.oracle.wls.exporter;
+package com.oracle.wls.exporter.sidecar;
 
 import java.util.concurrent.TimeUnit;
 
+import com.oracle.wls.exporter.WebClientFactoryImpl;
 import io.helidon.common.LogConfig;
 import io.helidon.webserver.Routing;
 import io.helidon.webserver.WebServer;
