+++ wls-exporter-sidecar/src/main/java/com/oracle/wls/exporter/sidecar/SidecarConfiguration.java
@@ -1,7 +1,7 @@
 // Copyright (c) 2021, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package com.oracle.wls.exporter;
+package com.oracle.wls.exporter.sidecar;
 
 public class SidecarConfiguration {
