+++ wls-exporter-sidecar/src/test/java/com/oracle/wls/exporter/sidecar/ItemHolder.java
@@ -1,7 +1,7 @@
 // Copyright (c) 2017, 2020, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package com.oracle.wls.exporter;
+package com.oracle.wls.exporter.sidecar;
 
 import java.util.ArrayList;
 import java.util.Arrays;
