+++ wls-exporter-sidecar/src/test/java/com/oracle/wls/exporter/sidecar/SidecarConfigurationTest.java
@@ -1,7 +1,7 @@
 // Copyright (c) 2021, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package com.oracle.wls.exporter;
+package com.oracle.wls.exporter.sidecar;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,12 +12,12 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static com.oracle.wls.exporter.SidecarConfiguration.DEFAULT_LISTEN_PORT;
-import static com.oracle.wls.exporter.SidecarConfiguration.DEFAULT_POD_NAME;
-import static com.oracle.wls.exporter.SidecarConfiguration.DEFAULT_WLS_PORT;
-import static com.oracle.wls.exporter.SidecarConfiguration.LISTEN_PORT_PROPERTY;
-import static com.oracle.wls.exporter.SidecarConfiguration.POD_NAME_PROPERTY;
-import static com.oracle.wls.exporter.SidecarConfiguration.WLS_PORT_PROPERTY;
+import static com.oracle.wls.exporter.sidecar.SidecarConfiguration.DEFAULT_LISTEN_PORT;
+import static com.oracle.wls.exporter.sidecar.SidecarConfiguration.DEFAULT_POD_NAME;
+import static com.oracle.wls.exporter.sidecar.SidecarConfiguration.DEFAULT_WLS_PORT;
+import static com.oracle.wls.exporter.sidecar.SidecarConfiguration.LISTEN_PORT_PROPERTY;
+import static com.oracle.wls.exporter.sidecar.SidecarConfiguration.POD_NAME_PROPERTY;
+import static com.oracle.wls.exporter.sidecar.SidecarConfiguration.WLS_PORT_PROPERTY;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
